changed argument for list deployment
