+++ cmd/list.go
@@ -1,5 +1,5 @@
 /*
-Copyright © 2019 NAME HERE <EMAIL ADDRESS>
+Copyright © 2019 Tomokazu HIRAI <tomokazu.hirai@gmail.com>
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,17 +16,14 @@ limitations under the License.
 package cmd
 
 import (
-	"flag"
 	"fmt"
-	"path/filepath"
 
 	"github.com/spf13/cobra"
-	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/util/homedir"
 )
+
+var namespace string
 
 // listCmd represents the list command
 var listCmd = &cobra.Command{
@@ -50,33 +47,7 @@ For example:
 
 minikubectl list deployment.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		var kubeconfig *string
-		if home := homedir.HomeDir(); home != "" {
-			kubeconfig = flag.String("kubeconfig", filepath.Join(home, ".kube", "config"), "(optional) absolute path to the kubeconfig file")
-		} else {
-			kubeconfig = flag.String("kubeconfig", "", "absolute path to the kubeconfig file")
-		}
-		flag.Parse()
-
-		config, err := clientcmd.BuildConfigFromFlags("", *kubeconfig)
-		if err != nil {
-			panic(err)
-		}
-
-		clientset, err := kubernetes.NewForConfig(config)
-		if err != nil {
-			panic(err)
-		}
-		deploymentsClient := clientset.AppsV1().Deployments(apiv1.NamespaceDefault)
-		list, err := deploymentsClient.List(metav1.ListOptions{})
-		if err != nil {
-			panic(err)
-		}
-		fmt.Printf("🍺 There are %d deployments in the cluster\n", len(list.Items))
-		for _, d := range list.Items {
-			fmt.Printf(" * %s (%d replicas)\n", d.Name, *d.Spec.Replicas)
-		}
-
+		listDeployments()
 	},
 }
 
@@ -89,32 +60,7 @@ For example:
 
 minikubectl list pod.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		var kubeconfig *string
-		if home := homedir.HomeDir(); home != "" {
-			kubeconfig = flag.String("kubeconfig", filepath.Join(home, ".kube", "config"), "(optional) absolute path to the kubeconfig file")
-		} else {
-			kubeconfig = flag.String("kubeconfig", "", "absolute path to the kubeconfig file")
-		}
-		flag.Parse()
-
-		config, err := clientcmd.BuildConfigFromFlags("", *kubeconfig)
-		if err != nil {
-			panic(err)
-		}
-
-		clientset, err := kubernetes.NewForConfig(config)
-		if err != nil {
-			panic(err)
-		}
-
-		pods, err := clientset.CoreV1().Pods("").List(metav1.ListOptions{})
-		if err != nil {
-			panic(err)
-		}
-		fmt.Printf("🍉 There are %d pods in the cluster\n", len(pods.Items))
-		for _, d := range pods.Items {
-			fmt.Printf(" * %s\n", d.Name)
-		}
+		listPods()
 	},
 }
 
@@ -122,20 +68,16 @@ func init() {
 	rootCmd.AddCommand(listCmd)
 	listCmd.AddCommand(listDeploymentCmd)
 	listCmd.AddCommand(listPodCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// listCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// listCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	listPodCmd.Flags().StringVarP(&namespace, "namespace", "n", "", "namespace name")
+	listDeploymentCmd.Flags().StringVarP(&namespace, "namespace", "n", "", "namespace name")
 }
 
-func listDeployments(clientset *kubernetes.Clientset) {
-	deploymentsClient := clientset.AppsV1().Deployments(apiv1.NamespaceDefault)
+func listDeployments() {
+	config := loadConfig()
+
+	clientset, err := kubernetes.NewForConfig(config)
+
+	deploymentsClient := clientset.AppsV1().Deployments(namespace)
 	list, err := deploymentsClient.List(metav1.ListOptions{})
 	if err != nil {
 		panic(err)
@@ -146,8 +88,12 @@ func listDeployments(clientset *kubernetes.Clientset) {
 	}
 }
 
-func listPods(clientset *kubernetes.Clientset) {
-	pods, err := clientset.CoreV1().Pods("").List(metav1.ListOptions{})
+func listPods() {
+	config := loadConfig()
+
+	clientset, err := kubernetes.NewForConfig(config)
+
+	pods, err := clientset.CoreV1().Pods(namespace).List(metav1.ListOptions{})
 	if err != nil {
 		panic(err)
 	}
